requirements.txt
fastapi
uvicorn
pytube
aiofiles # static files serving
openai
python-dotenv